# app/routes/admin.py
import os
import re
from datetime import datetime
from typing import Literal
from uuid import UUID

import asyncpg
from fastapi import APIRouter, Depends, HTTPException, Query
from pydantic import BaseModel, Field, field_validator

from ..auth import require_user
from ..db import get_conn
from ..migrations import db_alembic_revision, image_alembic_head

# Loose RFC-5322-shaped email check. Picks up obvious typos (no @,
# trailing space) without pulling in the email-validator package just
# for one form. The SPA's <input type="email"> covers the rest.
_EMAIL_RE = re.compile(r"^[^\s@]+@[^\s@]+\.[^\s@]+$")
_VALID_ROLES = ("consultant", "assistant", "admin")

router = APIRouter(prefix="/api/admin", tags=["admin"])


# ---- Response models -------------------------------------------------------

class AdminUser(BaseModel):
    id: UUID
    email: str
    name: str
    role: str
    is_active: bool
    last_login_at: datetime | None
    created_at: datetime


class CreateUserRequest(BaseModel):
    email: str = Field(min_length=3, max_length=320)
    name: str = Field(min_length=1, max_length=200)
    role: Literal["consultant", "assistant", "admin"]

    @field_validator("email")
    @classmethod
    def _email_shape(cls, v: str) -> str:
        v = v.strip()
        if not _EMAIL_RE.match(v):
            raise ValueError("must be a valid email address")
        return v

    @field_validator("name")
    @classmethod
    def _name_strip(cls, v: str) -> str:
        return v.strip()


class AuditLogEntry(BaseModel):
    id: int
    ts: datetime
    user_id: UUID | None
    user_email: str | None
    user_name: str | None
    table_name: str
    row_id: UUID | None
    action: str


class AuditLogPage(BaseModel):
    items: list[AuditLogEntry]
    total: int
    limit: int
    offset: int


class AboutInfo(BaseModel):
    build_commit: str
    api_title: str
    counts: dict[str, int]
    db_revision: str | None
    image_head_revision: str | None
    migration_in_sync: bool


# ---- Routes ----------------------------------------------------------------

@router.get("/users", response_model=list[AdminUser])
async def list_users(_user=Depends(require_user), conn=Depends(get_conn)):
    """All users, sorted by name. Anyone authenticated can read this for
    now; if the role surface grows we'll gate this on an admin role."""
    rows = await conn.fetch(
        """
        SELECT id, email, name, role, is_active, last_login_at, created_at
        FROM users
        ORDER BY is_active DESC, name
        """
    )
    return [dict(r) for r in rows]


@router.post("/users", response_model=AdminUser, status_code=201)
async def create_user(
    body: CreateUserRequest,
    _user=Depends(require_user),
    conn=Depends(get_conn),
):
    """Pre-authorize a new user. Until they sign in via Google OAuth
    the row exists with last_login_at NULL; their first successful
    OAuth callback flips that. Email is CITEXT UNIQUE — duplicates
    return 409. Role gating for THIS endpoint is intentionally not
    enforced yet (matches list_users); the same caveat applies."""
    try:
        row = await conn.fetchrow(
            """
            INSERT INTO users (email, name, role, is_active)
            VALUES ($1, $2, $3, TRUE)
            RETURNING id, email, name, role, is_active, last_login_at, created_at
            """,
            body.email, body.name, body.role,
        )
    except asyncpg.UniqueViolationError as e:
        raise HTTPException(status_code=409, detail="A user with that email already exists.") from e
    return dict(row)


@router.delete("/users/{user_id}", response_model=AdminUser)
async def deactivate_user(
    user_id: UUID,
    user=Depends(require_user),
    conn=Depends(get_conn),
):
    """Soft-delete: set is_active=false. Preserves audit_log.user_id
    foreign keys and lets us reactivate later. Refuses to deactivate
    the caller's own account so an admin can't lock themselves out."""
    if user and user["id"] == user_id:
        raise HTTPException(status_code=400, detail="You cannot deactivate your own account.")
    row = await conn.fetchrow(
        """
        UPDATE users
           SET is_active = FALSE, updated_at = NOW()
         WHERE id = $1
        RETURNING id, email, name, role, is_active, last_login_at, created_at
        """,
        user_id,
    )
    if row is None:
        raise HTTPException(status_code=404, detail="User not found.")
    return dict(row)


@router.post("/users/{user_id}/reactivate", response_model=AdminUser)
async def reactivate_user(
    user_id: UUID,
    _user=Depends(require_user),
    conn=Depends(get_conn),
):
    """Reverse of deactivate_user. Idempotent: reactivating an already-
    active user is a no-op that still returns the current row."""
    row = await conn.fetchrow(
        """
        UPDATE users
           SET is_active = TRUE, updated_at = NOW()
         WHERE id = $1
        RETURNING id, email, name, role, is_active, last_login_at, created_at
        """,
        user_id,
    )
    if row is None:
        raise HTTPException(status_code=404, detail="User not found.")
    return dict(row)


@router.get("/audit-log", response_model=AuditLogPage)
async def list_audit_log(
    limit: int = Query(50, ge=1, le=200),
    offset: int = Query(0, ge=0),
    _user=Depends(require_user),
    conn=Depends(get_conn),
):
    """Recent audit_log entries, newest first. Offset-paginated;
    fine for the homelab scale (a few thousand rows). If we hit the
    wall, switch to keyset on audit_log.id."""
    total = await conn.fetchval("SELECT COUNT(*) FROM audit_log")
    rows = await conn.fetch(
        """
        SELECT al.id, al.ts, al.user_id, u.email AS user_email, u.name AS user_name,
               al.table_name, al.row_id, al.action
        FROM audit_log al
        LEFT JOIN users u ON u.id = al.user_id
        ORDER BY al.id DESC
        LIMIT $1 OFFSET $2
        """,
        limit, offset,
    )
    return {
        "items": [dict(r) for r in rows],
        "total": total or 0,
        "limit": limit,
        "offset": offset,
    }


@router.get("/about", response_model=AboutInfo)
async def about(_user=Depends(require_user), conn=Depends(get_conn)):
    """Build/runtime metadata for the about/settings tab. Counts give a
    quick sanity check on whether the app sees the data it should."""
    counts_row = await conn.fetchrow(
        """
        SELECT
          (SELECT COUNT(*) FROM users)               AS users,
          (SELECT COUNT(*) FROM families
             WHERE deleted_at IS NULL)               AS families,
          (SELECT COUNT(*) FROM people
             WHERE kind = 'student' AND deleted_at IS NULL) AS students,
          (SELECT COUNT(*) FROM engagements
             WHERE deleted_at IS NULL)               AS engagements,
          (SELECT COUNT(*) FROM schools
             WHERE deleted_at IS NULL)               AS schools,
          (SELECT COUNT(*) FROM service_items
             WHERE deleted_at IS NULL)               AS service_items,
          (SELECT COUNT(*) FROM audit_log)           AS audit_log_entries
        """
    )
    db_rev = await db_alembic_revision(conn)
    img_head = image_alembic_head()
    return {
        "build_commit": os.environ.get("BUILD_COMMIT", "dev"),
        "api_title": "HillCo2 API",
        "counts": dict(counts_row) if counts_row else {},
        "db_revision": db_rev,
        "image_head_revision": img_head,
        "migration_in_sync": (
            db_rev is not None and img_head is not None and db_rev == img_head
        ),
    }